Install node modules in build
- Detect requests node for launch and build
- Detect requests modules for launch
- Install Pack cli off master branch
- Add acceptance test that uses node-cnb-buildpack from git release

[#160879447]

+++ scripts/install_tools.sh
@@ -10,4 +10,11 @@ fi
 export GOBIN=$PWD/.bin
 export PATH=$GOBIN:$PATH
 
-go install github.com/onsi/ginkgo/ginkgo
+if [[ ! -f $GOBIN/ginkgo ]]; then
+    go install github.com/onsi/ginkgo/ginkgo
+fi
+
+if [[ ! -f $GOBIN/pack ]]; then
+    go get github.com/buildpack/pack@master
+    go install github.com/buildpack/pack/cmd/pack
+fi
